requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy==1.25.2
tqdm==4.66.1
tensorboard==2.14.1
matplotlib==3.8.0
pandas==2.1.1
audiomentations==0.33.0
speechbrain~=0.5.12
datasets==2.14.5
torch_audiomentations==0.11.0
wandb==0.15.12
torchaudio==2.0.2
pillow
librosa==0.10.1
transformers==4.35.0
tokenizers==0.14.1
gdown==4.7.1
pyloudnorm==0.1.1
accelerate==0.24.1
torchmetrics==0.11.4
pesq==0.0.4